src/requirements.txt
azure-mgmt-resource
crhelper
azure-identity
